src/main/java/de/flingelli/security/openvas/reports/data/Scanner.java
package de.flingelli.security.openvas.reports.data;

import javax.xml.bind.annotation.XmlAccessType;
import javax.xml.bind.annotation.XmlAccessorType;
import javax.xml.bind.annotation.XmlAttribute;
import javax.xml.bind.annotation.XmlElement;

@XmlAccessorType(XmlAccessType.FIELD)
public class Scanner {
    @XmlAttribute
    private String id;
    @XmlElement
    private String name;
    @XmlElement
    private int type;
    @XmlElement
    private int trash;

    public String getId() {
        return id;
    }

    public String getName() {
        return name;
    }

    public int getType() {
        return type;
    }

    public int getTrash() {
        return trash;
    }
}
